Match Math.sqrt to default behaviour with Integers
I realise that the mathn library changes this, but I think most developers rely on Math.sqrt returning a float like the default math.rb library. As such when we use ruby-unit + other gems that rely on Math.sqrt returning a Float we can get issues (alchemist, or rgeo for example).