Replace fixed class method calls with self(.class) in Unit class
Replaces every class method call `RubyUnits::Unit.method` with either `method`
(in class methods) or `self.class.method` (in instance methods).

This allows us to override a class method in a sub class and make sure that
every call to that method from it's instances use the new definition.

E.g. we can have a sub class which extends from RubyUnits::Unit and override
eliminate_terms which is then used when dividing or multiplicating instances
of this sub class.